Changed the "audition" function into an extension of the AVAudioPCMBuffer class itself, which feels a bit more Swifty, removes the need for a parameter (since "self" becomes the implicit parameter), and avoids polluting the global function namespace.
